requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
Flask>=3.0.0 